Make sure we don't load detect-indent (not compatible with old node versions) when only testing postinstall
